+++ lib/surface/components/form/field_context.ex
@@ -12,7 +12,7 @@ defmodule Surface.Components.Form.FieldContext do
   alias Surface.Components.Form.Field
 
   @doc "The field name"
-  prop name, :atom, required: true
+  prop name, :any, required: true
 
   @doc """
   The content for the field